If scrollbar's detached, reinitialise the plugin in update.
