+++ src/js/plugin/handler/click-rail.js
@@ -5,7 +5,7 @@
 
 var h = require('../../lib/helper')
   , instances = require('../instances')
-  , updateGeometry = require('../update');
+  , updateGeometry = require('../update-geometry');
 
 function bindClickRailHandler(element, i) {
   function pageOffset(el) {
