+++ src/js/plugin/handler/mouse-wheel.js
@@ -5,7 +5,7 @@
 
 var h = require('../../lib/helper')
   , instances = require('../instances')
-  , updateGeometry = require('../update');
+  , updateGeometry = require('../update-geometry');
 
 function bindMouseWheelHandler(element, i) {
   var shouldPrevent = false;
