+++ src/js/plugin/handler/native-scroll.js
@@ -4,7 +4,7 @@
 'use strict';
 
 var instances = require('../instances')
-  , updateGeometry = require('../update');
+  , updateGeometry = require('../update-geometry');
 
 function bindNativeScrollHandler(element, i) {
   i.event.bind(element, 'scroll', function () {
