+++ src/js/plugin/handler/selection.js
@@ -5,7 +5,7 @@
 
 var h = require('../../lib/helper')
   , instances = require('../instances')
-  , updateGeometry = require('../update');
+  , updateGeometry = require('../update-geometry');
 
 function bindSelectionHandler(element, i) {
   function getRangeNode() {
